Tests/LibWeb/Text/expected/DOM/CharacterData-replaceData-break-surrogate-pair.txt

Before replaceData:
[0]: 53997
[1]: 56998
After replaceData:
[0]: 56998
